+++ .formatter.exs
@@ -69,6 +69,7 @@ spark_locals_without_parens = [
   magic_link: 1,
   magic_link: 2,
   monitor_fields: 1,
+  multitenancy_relationship: 1,
   name: 1,
   nonce: 1,
   oauth2: 0,