+++ lib/ash_authentication/errors/invalid_token.ex
@@ -2,7 +2,13 @@ defmodule AshAuthentication.Errors.InvalidToken do
   @moduledoc """
   An invalid token was presented.
   """
-  use Splode.Error, fields: [:type, :field], class: :forbidden
+  use Splode.Error, fields: [:type, :field, :reason], class: :forbidden
 
-  def message(%{type: type}), do: "Invalid #{type} token"
+  def message(%{type: type, reason: reason}) do
+    if reason do
+      "Invalid #{type} token: #{reason}"
+    else
+      "Invalid #{type} token"
+    end
+  end
 end