+++ lib/ash_authentication/strategies/api_key.ex
@@ -27,7 +27,8 @@ defmodule AshAuthentication.Strategy.ApiKey do
             resource: nil,
             sign_in_action_name: nil,
             api_key_hash_attribute: nil,
-            api_key_relationship: nil
+            api_key_relationship: nil,
+            multitenancy_relationship: nil
 
   alias AshAuthentication.Strategy.{ApiKey, ApiKey.Transformer, ApiKey.Verifier}
 
@@ -38,7 +39,8 @@ defmodule AshAuthentication.Strategy.ApiKey do
           resource: Ash.Resource.t(),
           sign_in_action_name: atom(),
           api_key_hash_attribute: atom(),
-          api_key_relationship: atom()
+          api_key_relationship: atom(),
+          multitenancy_relationship: atom()
         }
 
   @doc false